feat[spotifymusic::banda]: banda service - criacao do servico - criacao endpoints - configuracao efcore para exclusao em cascata

+++ src/Applications/AVS.SpotifyMusic.Application/AppServices/UsuarioAppService.cs
@@ -118,7 +118,15 @@ namespace AVS.SpotifyMusic.Application.AppServices
 
 		public async Task<bool> Remover(Guid id)
 		{
-			var response = await _usuarioService.Remover(id);
+			var usuario = await _usuarioService.BuscarPorCriterioDetalhado(x => x.Id == id);
+			if (usuario == null) return false;
+            
+			usuario.LimparCartoes();
+            usuario.LimparAssinaturas();
+            usuario.LimparPlaylists();
+            usuario.LimparNotificacao();            
+
+            var response = await _usuarioService.Remover(id);
 			return response;
 		}
